Move post process hooks temporal logic
