+++ front/lib/extract_events.ts
@@ -1,10 +1,5 @@
 import { Op } from "sequelize";
 
-import type {
-  DocumentsPostProcessHookFilterParams,
-  DocumentsPostProcessHookOnUpsertParams,
-} from "@app/documents_post_process_hooks/hooks";
-import { getDatasource } from "@app/documents_post_process_hooks/hooks/lib/data_source_helpers";
 import type { Authenticator } from "@app/lib/auth";
 import {
   _getMaxTextContentToProcess,
@@ -20,6 +15,11 @@ import { EventSchema, ExtractedEvent } from "@app/lib/models";
 import { generateModelSId } from "@app/lib/utils";
 import logger from "@app/logger/logger";
 import { logOnSlack } from "@app/logger/slack_debug_logger";
+import type {
+  DocumentsPostProcessHookFilterParams,
+  DocumentsPostProcessHookOnUpsertParams,
+} from "@app/temporal/documents_post_process_hooks/hooks";
+import { getDatasource } from "@app/temporal/documents_post_process_hooks/hooks/lib/data_source_helpers";
 
 const { URL } = process.env;
