+++ front/start_worker.ts
@@ -1,8 +1,8 @@
 import { setupGlobalErrorHandler } from "@dust-tt/types";
 
-import { runPostUpsertHooksWorker } from "@app/documents_post_process_hooks/temporal/worker";
 import logger from "@app/logger/logger";
 import { runPokeWorker } from "@app/poke/temporal/worker";
+import { runPostUpsertHooksWorker } from "@app/temporal/documents_post_process_hooks/worker";
 import { runProductionChecksWorker } from "@app/temporal/production_checks/worker";
 import { runScrubWorkspaceQueueWorker } from "@app/temporal/scrub_workspace/worker";
 import { runUpsertQueueWorker } from "@app/temporal/upsert_queue/worker";
