+++ front/temporal/documents_post_process_hooks/hooks/document_tracker/suggest_changes/lib.ts
@@ -3,14 +3,6 @@ import sgMail from "@sendgrid/mail";
 import { Op } from "sequelize";
 import showdown from "showdown";
 
-import type {
-  DocumentsPostProcessHookFilterParams,
-  DocumentsPostProcessHookOnUpsertParams,
-} from "@app/documents_post_process_hooks/hooks";
-import {
-  getDatasource,
-  getDocumentDiff,
-} from "@app/documents_post_process_hooks/hooks/lib/data_source_helpers";
 import {
   DataSource,
   DocumentTrackerChangeSuggestion,
@@ -19,6 +11,14 @@ import {
   Workspace,
 } from "@app/lib/models";
 import mainLogger from "@app/logger/logger";
+import type {
+  DocumentsPostProcessHookFilterParams,
+  DocumentsPostProcessHookOnUpsertParams,
+} from "@app/temporal/documents_post_process_hooks/hooks";
+import {
+  getDatasource,
+  getDocumentDiff,
+} from "@app/temporal/documents_post_process_hooks/hooks/lib/data_source_helpers";
 
 import { callDocTrackerRetrievalAction } from "./actions/doc_tracker_retrieval";
 import { callDocTrackerSuggestChangesAction } from "./actions/doc_tracker_suggest_changes";
