+++ front/temporal/documents_post_process_hooks/hooks/index.ts
@@ -1,12 +1,12 @@
 import type { ConnectorProvider, UpsertContext } from "@dust-tt/types";
 
-import { DEFAULT_DOCUMENTS_POST_PROCESS_HOOKS_DEBOUNCE_MS } from "@app/documents_post_process_hooks/hooks/consts";
+import type { Authenticator } from "@app/lib/auth";
+import { DEFAULT_DOCUMENTS_POST_PROCESS_HOOKS_DEBOUNCE_MS } from "@app/temporal/documents_post_process_hooks/hooks/consts";
 import {
   documentTrackerSuggestChangesPostProcessHook,
   documentTrackerUpdateTrackedDocumentsPostProcessHook,
-} from "@app/documents_post_process_hooks/hooks/document_tracker";
-import { extractEventPostProcessHook } from "@app/documents_post_process_hooks/hooks/extract_event";
-import type { Authenticator } from "@app/lib/auth";
+} from "@app/temporal/documents_post_process_hooks/hooks/document_tracker";
+import { extractEventPostProcessHook } from "@app/temporal/documents_post_process_hooks/hooks/extract_event";
 
 export const DOCUMENTS_POST_PROCESS_HOOK_TYPES = [
   "document_tracker_update_tracked_documents",
