+++ front/temporal/documents_post_process_hooks/worker.ts
@@ -1,10 +1,10 @@
 import type { Context } from "@temporalio/activity";
 import { Worker } from "@temporalio/worker";
 
-import * as activities from "@app/documents_post_process_hooks/temporal/activities";
 import { getTemporalWorkerConnection } from "@app/lib/temporal";
 import { ActivityInboundLogInterceptor } from "@app/lib/temporal_monitoring";
 import logger from "@app/logger/logger";
+import * as activities from "@app/temporal/documents_post_process_hooks/activities";
 
 export async function runPostUpsertHooksWorker() {
   const { connection, namespace } = await getTemporalWorkerConnection();
